タスク15完了

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -44,14 +44,14 @@ public class DeleteBookController {
 			Model model) {
 		logger.info("Welcome delete! The client locale is {}.", locale);
 
-		if (rentBooksService.getRentBookInfo(bookId) != 0) {
+		if (booksService.isCheckedOut(bookId)) {
 			model.addAttribute("rentErrorMessage", "貸出中の本は削除できません");
 			model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
 			return "details";
-    } else {
-      booksService.deleteBook(bookId);
+		} else {
+			booksService.deleteBook(bookId);
+			return "redirect:home";
     }
-		return "redirect:home";
 	}
 
 }
